requirements.txt
numpy>=1.26
scipy>=1.13
matplotlib>=3.8
pydantic>=2.0
hydra-core>=1.3
networkx>=3.2
shapely>=2.0
pyyaml>=6.0
rich>=13.0